[MINOR] Remove CacheObjectContext dependency from binary
