+++ modules/core/src/main/java/org/apache/ignite/internal/processors/cache/CacheObject.java
@@ -67,7 +67,7 @@ public interface CacheObject extends Message {
      * @return Size required to store this value object.
      * @throws IgniteCheckedException If failed.
      */
-    public int valueBytesLength(CacheObjectContext ctx) throws IgniteCheckedException;
+    public int valueBytesLength(CacheObjectValueContext ctx) throws IgniteCheckedException;
 
     /**
      * @param buf Buffer to write value to.
@@ -114,7 +114,7 @@ public interface CacheObject extends Message {
      * @param ctx Cache context.
      * @return Instance to store in cache.
      */
-    public CacheObject prepareForCache(CacheObjectContext ctx);
+    public CacheObject prepareForCache(CacheObjectValueContext ctx);
 
     /**
      * @param ctx Context.
